--- app-integration/src/main/java/com/example/learning/LearningApplication.java
@@ -1,12 +1,0 @@
-package com.example.learning;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LearningApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(LearningApplication.class, args);
-    }
-}
