--- app-integration/src/main/java/com/example/learning/controller/ExternalOrderProcessor.java
@@ -1,11 +1,0 @@
-package com.example.learning.controller;
-
-import org.springframework.stereotype.Component;
-@Component
-public class ExternalOrderProcessor implements OrderProcesser {
-
-    public String test() {
-
-        return this.getClass().getName();
-    }
-}
