--- app-integration/src/main/java/com/example/learning/controller/OrderProcesser.java
@@ -1,6 +1,0 @@
-package com.example.learning.controller;
-
-public interface OrderProcesser {
-
-    String test();
-}
