--- app-integration/src/main/java/com/example/learning/entity/Order.java
@@ -1,11 +1,0 @@
-package com.example.learning.entity;
-
-import lombok.Builder;
-import lombok.Data;
-@Builder
-@Data
-public class Order {
-
-    private String name;
-    private int quantity;
-}
