--- app-integration/src/main/java/com/example/learning/kafka/exceptions/CustomException.java
@@ -1,4 +1,0 @@
-package com.example.learning.kafka.exceptions;
-
-public class CustomException extends Exception {
-}
