--- app-integration/src/main/java/com/example/learning/kafka/producer/AppConstants.java
@@ -1,7 +1,0 @@
-package com.example.learning.kafka.producer;
-
-public class AppConstants {
-    public static final String TOPIC_JSON = "producer_json";
-    public static final String TOPIC_STRING = "producer_string";
-    public static final String GROUP_ID = "group_id";
-}
