--- app-integration/src/main/java/com/example/learning/kafka/producer/Order.java
@@ -1,21 +1,0 @@
-package com.example.learning.kafka.producer;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Builder
-@ToString
-@RequiredArgsConstructor
-@AllArgsConstructor
-@Setter
-@Getter
-public class Order {
-
-    private String id;
-
-    private String name;
-}
