+++ app-integration/src/main/java/com/shopping/LearningApplication.java
@@ -1,0 +1,12 @@
+package com.shopping;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LearningApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(LearningApplication.class, args);
+    }
+}
