+++ app-integration/src/main/java/com/shopping/controller/InternalOrderProcessor.java
@@ -1,0 +1,11 @@
+package com.shopping.controller;
+
+import org.springframework.stereotype.Component;
+@Component
+public class InternalOrderProcessor implements OrderProcesser {
+
+    public String test() {
+
+        return this.getClass().getName();
+    }
+}
