+++ app-integration/src/main/java/com/shopping/controller/OrderProcesser.java
@@ -1,0 +1,6 @@
+package com.shopping.controller;
+
+public interface OrderProcesser {
+
+    String test();
+}
