+++ app-integration/src/main/java/com/shopping/entity/Order.java
@@ -1,0 +1,11 @@
+package com.shopping.entity;
+
+import lombok.Builder;
+import lombok.Data;
+@Builder
+@Data
+public class Order {
+
+    private String name;
+    private int quantity;
+}
