+++ app-integration/src/main/java/com/shopping/kafka/exceptions/CustomException.java
@@ -1,0 +1,4 @@
+package com.shopping.kafka.exceptions;
+
+public class CustomException extends Exception {
+}
