+++ app-integration/src/main/java/com/shopping/kafka/producer/AppConstants.java
@@ -1,0 +1,7 @@
+package com.shopping.kafka.producer;
+
+public class AppConstants {
+    public static final String TOPIC_JSON = "producer_json";
+    public static final String TOPIC_STRING = "producer_string";
+    public static final String GROUP_ID = "group_id";
+}
