--- app-integration/src/test/java/com/example/learning/BagCucumberIntegrationTest.java
@@ -1,11 +1,0 @@
-package com.example.learning;
-
-import org.junit.runner.RunWith;
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
-@RunWith(Cucumber.class)
-@CucumberOptions(features = "classpath:features",
-        plugin = {"pretty", "html:target/cucumber/report.html"}
-)
-public class BagCucumberIntegrationTest {
-}
