--- app-integration/src/test/java/com/example/learning/LearningApplicationTests.java
@@ -1,15 +1,0 @@
-package com.example.learning;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import io.cucumber.spring.CucumberContextConfiguration;
-
-@CucumberContextConfiguration
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class LearningApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
