--- app-integration/src/test/java/com/example/learning/steps/Stepdefinitions.java
@@ -1,34 +1,0 @@
-package com.example.learning.steps;
-
-import static org.junit.Assert.assertEquals;
-
-import com.example.learning.vaidation.Validator;
-import io.cucumber.java8.En;
-
-public class Stepdefinitions implements En {
-
-    String today;
-    String actualAnswer;
-
-    public Stepdefinitions() {
-
-        Given("^today is \"([^\"]*)\"$", (String day) -> {
-
-            today = day;
-        });
-
-        When(
-                "I ask whether it's Friday yet",
-                () -> {
-                    actualAnswer = Validator.validate(today);
-                });
-        Then(
-                "I should be told {string}",
-                (String resultAnswer) -> {
-                    assertEquals(actualAnswer, resultAnswer);
-                });
-        Given("^today is Sunday$", () -> {
-            today = "Sunday";
-        });
-    }
-}
