+++ app-integration/src/test/java/com/shopping/BagCucumberIntegrationTest.java
@@ -1,0 +1,11 @@
+package com.shopping;
+
+import org.junit.runner.RunWith;
+import io.cucumber.junit.Cucumber;
+import io.cucumber.junit.CucumberOptions;
+@RunWith(Cucumber.class)
+@CucumberOptions(features = "classpath:features",
+        plugin = {"pretty", "html:target/cucumber/report.html"}
+)
+public class BagCucumberIntegrationTest {
+}
