+++ app-integration/src/test/java/com/shopping/LearningApplicationTests.java
@@ -1,0 +1,15 @@
+package com.shopping;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import io.cucumber.spring.CucumberContextConfiguration;
+
+@CucumberContextConfiguration
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class LearningApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
